LayoutTests/platform/chromium-mac/fast/transforms/selection-bounds-in-transformed-view-expected.txt

CONSOLE MESSAGE: line 10: Uncaught TypeError: Object [object Object] has no method 'scalePageBy'
target
